pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "memos"
version = "0.4.4"
description = "A package for memos"
readme = "README.md"
authors = [{ name = "arkohut" }]
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "pydantic",
    "sqlalchemy",
    "typer",
    "magika",
    "pydantic-settings",
    "typesense",
    "opencv-python",
    "pillow",
    "piexif",
    "imagehash",
    "screeninfo",
    "pywin32; sys_platform == 'win32'",
    "psutil; sys_platform == 'win32'",
    "pyobjc-framework-Quartz; sys_platform == 'darwin'",
    "pyobjc-framework-AppKit; sys_platform == 'darwin'",
]

[project.urls]
Homepage = "https://github.com/arkohut/memos"

[project.scripts]
memos = "memos.commands:app"

[tool.setuptools.packages.find]
include = ["memos*", "screen_recorder*"]

[tool.setuptools.package-data]
"*" = ["static/**/*"]
